% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emiss_functions.R
\name{assign_cardinal}
\alias{assign_cardinal}
\title{Assign cardinal direction to wind direction.}
\usage{
assign_cardinal(df1, dir_column)
}
\arguments{
\item{df1}{Data frame. Data frame containing wind direction (in angle).}

\item{dir_column.}{Integer. Number of wind direction column in df1.}
}
\value{
Data frame. Original data frame df1 with additional column listing 
named wind direction.
}
\description{
Go through a data file containing wind directions (in angles) and assign one 
of the 16 cardinal directions (in a new column).
}
\examples{
assign_cardinal(mydata, 3)
}

