% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emiss_functions.R
\name{direction_reference}
\alias{direction_reference}
\title{Build reference list for cardinal direction angles.}
\usage{
direction_reference()
}
\value{
A named list of the 16 cardinal directions and their angle ranges.
}
\description{
Build reference list for cardinal direction angles.
}

