% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emiss_functions.R
\name{multi_read}
\alias{multi_read}
\title{Read in multiple data files}
\usage{
multi_read(path, filenames)
}
\arguments{
\item{path}{String. Path for data files.}

\item{filenames}{String. File names. Can be a vector for multiple files.}
}
\value{
Data frame.
}
\description{
Read in multiple files of the same format and append all into one data frame.
Additional column is added with name of source data file for each row.
}
\examples{
multi_read("./data-raw/", c("file1.csv", "file2.csv", "file3.csv"))
}

