% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rose_functions.R
\name{plot_identity_rose}
\alias{plot_identity_rose}
\title{Create directional rose ggplot object}
\usage{
plot_identity_rose(data, value, dir, valuemin = 0, valueseq,
  palette = "YlOrRd", legend.title, plot.title, ylabel.loc = "NW")
}
\arguments{
\item{data}{Data frame.}

\item{value}{Text string. Column name of value variable.}

\item{dir}{Text string. Column name of direction variable.}

\item{valuemin}{Numeric. Lower value limit. All rows with value < valuemin 
are removed from data.}

\item{valueseq}{Numberic. Bin cutoffs for values.}

\item{palette}{Text string. Name of ColorBrewer palette to use for value 
fill.}

\item{legend.title}{Text string.}

\item{plot.title}{Text string.}
}
\value{
Returns the plot as a ggplot object. Paddles lengths (total and fill)
are proportional to percentage of total.
}
\description{
rose based on direct identiy data assopciated with direction.
}
\examples{
plot_identity_rose(datafile, value='sand_flux', dir='resultant_wd_10m',
          valuemin=low.cutoff, legend.title="Sand Flux",
          plot.title="Sand Roses")
}

